+++ system/src/main/java/io/github/sds100/keymapper/system/volume/RingerMode.kt
@@ -1,5 +1,8 @@
 package io.github.sds100.keymapper.system.volume
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 enum class RingerMode {
     NORMAL,
     VIBRATE,